Fix broken compilation in linux
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3203 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/comscore/main_samson_comscore_directory_create.cpp
@@ -200,7 +200,7 @@ void read_original_dictionary_file( const char * file_name , size_t max_num_reco
     while( fgets( line, 10000 , file ) )
     {
         fields[0] = line;
-        for ( int f = 1 ; f < (sizeof(fields)/sizeof(char*)) ; f++ )
+        for ( int f = 1 ; f < (int)(sizeof(fields)/sizeof(char*)) ; f++ )
         {
             char * pos = strstr( fields[f-1], "\t" );
             *pos = '\0';
